+++ AE-go_GameServer/src/com/aionemu/gameserver/model/drop/DropItem.java
@@ -27,7 +27,7 @@ public class DropItem
 
 	private int index = 0;
 
-	private int count = 0;
+	private long count = 0;
 
 	private DropTemplate dropTemplate;
 
@@ -70,7 +70,7 @@ public class DropItem
 	/**
 	 * @return the count
 	 */
-	public int getCount()
+	public long getCount()
 	{
 		return count;
 	}
@@ -78,7 +78,7 @@ public class DropItem
 	/**
 	 * @param count
 	 */
-	public void setCount(int count)
+	public void setCount(long count)
 	{
 		this.count = count;
 	}
