+++ AE-go_GameServer/src/com/aionemu/gameserver/model/trade/TradePSItem.java
@@ -23,13 +23,13 @@ package com.aionemu.gameserver.model.trade;
 public class TradePSItem extends TradeItem
 {
 	private int	itemObjId;
-	private int	price;
+	private long	price;
 
 	/**
 	 * @param itemId
 	 * @param count
 	 */
-	public TradePSItem(int itemObjId, int itemId, int count, int price)
+	public TradePSItem(int itemObjId, int itemId, long count, long price)
 	{
 		super(itemId, count);
 		this.setPrice(price);
@@ -40,7 +40,7 @@ public class TradePSItem extends TradeItem
 	 * @param price
 	 *            the price to set
 	 */
-	public void setPrice(int price)
+	public void setPrice(long price)
 	{
 		this.price = price;
 	}
@@ -48,7 +48,7 @@ public class TradePSItem extends TradeItem
 	/**
 	 * @return the price
 	 */
-	public int getPrice()
+	public long getPrice()
 	{
 		return price;
 	}
